src/engines/browser-use/server/requirements.txt
browser-use>=0.1.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
python-dotenv>=1.0.0
